test: extract processing unit tests
